Make --log argument case-insensitive
The `--log` command-line argument is now case-insensitive to provide a
better user experience.

Previously, the log level had to be specified in all caps (e.g.,
`INFO`). Inputs like `info` or `Debug` would cause an argument parsing
error.

This is resolved by setting `type=str.upper` in the argument's
definition, which normalizes the user's input before it is validated
against the list of choices.